microtar: strcpy -> strncpy
On mac, when extracting the source code tarball, there are some files
with paths longer than 100 characters that use a format from GNU tar
where a "././@LongLink" block is used to hold the full filepath and the
block that holds the file data has the path truncated to 100 characters.
microtar was seemingly overrunning its 100 char name buffers by copying
the full name in, ultimately causing crashes on mac. this commit fixes
the crashes by using strncpy instead of strcpy. This doesn't handle
@LongLink entries though, so these long file names will be truncated.
This only happens for wasm3 xcode project files on ios that have huge
path names, so it doesn't seem like a problem for us.