c-tooling: generate windows icon in bundle command
In order to link an icon into the executable, the runtime now has to be
shipped as an object file along with the static libs for wasm and orca
core. The bundle command generates a windows icon from the user supplied
image and links it with the runtime object files to generate the final
executable.