c-tooling: embed version in cli tool
Added a --version flag to the build-tool command, so that CI can embed
version strings like vX.X.X into the tool. Now when the update command
is called and the tool needs to check if it needs to replace itself,
this version can be checked instead of having to read a current_version
file.